refactor : notification폴더 global 안에서 다시 빼냄
Parent에서는 작동하는데 child에서 작동안함
