+++ server2/server-child/src/main/java/sinhan/server3/Server2Application.java
@@ -1,11 +1,15 @@
-package sinhan.server1;
+package sinhan.server3;
 
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import sinhan.server2.Server3Application;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"sinhan.server2"})
+@EnableJpaRepositories(basePackages = {"sinhan.server2.domain"})
+@EnableMongoRepositories(basePackages = {"sinhan.server2.notification.mongo"})
 public class Server2Application {
 
     public static void main(String[] args) {
