+++ server2/server-common/src/main/java/sinhan/server2/notification/controller/NotificationController.java
@@ -1,12 +1,12 @@
-package sinhan.server2.global.notification.controller;
+package sinhan.server2.notification.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-import sinhan.server2.global.notification.dto.NotificationFindAllResponse;
+import sinhan.server2.notification.dto.NotificationFindAllResponse;
 import sinhan.server2.global.utils.ApiUtils;
-import sinhan.server2.global.notification.service.SSEService;
+import sinhan.server2.notification.service.SSEService;
 
 import java.util.List;
