+++ server2/server-common/src/main/java/sinhan/server2/notification/entity/Notification.java
@@ -1,4 +1,4 @@
-package sinhan.server2.global.notification.entity;
+package sinhan.server2.notification.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
