+++ server2/server-common/src/main/java/sinhan/server2/notification/mongo/NotificationRepository.java
@@ -1,7 +1,7 @@
-package sinhan.server2.global.notification.mongo;
+package sinhan.server2.notification.mongo;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import sinhan.server2.global.notification.entity.Notification;
+import sinhan.server2.notification.entity.Notification;
 
 import java.util.List;
